Version of BaseControllerActivity that extends FragmentActivity.
BaseControllerActivity now extends FragmentActivity so that the FragmentActivity activity methods such as getSupportLoaderManager() are available to classes that extend BaseControllerActivity.

The onRetainNonConfigurationInstance() method is marked as final by FragmentActivity, so it had to be removed from the BaseControllerActivity and Controller classes. The alternative method provided by FragmentActivity, onRetainCustomNonConfigurationInstance(),  has been added instead. so it had to be removed from